backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pillow==11.2.1
numpy==2.2.5
torch==2.7.0
